+++ packages/ui/src/components/Alert/__stories__/Title.stories.tsx
@@ -3,10 +3,9 @@ import { Template } from './Template.stories'
 export const Title = Template.bind({})
 
 Title.args = {
-  icon: 'information-outline',
   variant: 'info',
   title: 'Information',
-  children: 'This is a notification bar with a custom title.',
+  children: 'This is an alert content.',
 }
 
 Title.parameters = {