+++ packages/ui/src/components/Alert/__tests__/index.test.tsx
@@ -1,5 +1,10 @@
+import { screen, waitFor } from '@testing-library/react'
+import userEvent from '@testing-library/user-event'
 import { Alert } from '..'
-import { shouldMatchEmotionSnapshot } from '../../../../.jest/helpers'
+import {
+  renderWithTheme,
+  shouldMatchEmotionSnapshot,
+} from '../../../../.jest/helpers'
 
 describe('Alert', () => {
   beforeAll(() => {
@@ -13,23 +18,54 @@ describe('Alert', () => {
   test('renders correctly with default values', () =>
     shouldMatchEmotionSnapshot(<Alert>Sample Alert</Alert>))
 
-  test('renders correctly with custom icon', () =>
-    shouldMatchEmotionSnapshot(<Alert icon="eye">Sample Alert</Alert>))
+  test('renders correctly with children as component', () =>
+    shouldMatchEmotionSnapshot(
+      <Alert>
+        <p>Sample Alert</p>
+      </Alert>,
+    ))
 
   test('renders correctly with title', () =>
     shouldMatchEmotionSnapshot(<Alert title="title">Sample Alert</Alert>))
 
-  test(`renders correctly unknown variant and fallback to standard`, () =>
-    // @ts-expect-error "unknown" isn't part of the `variant` variant, but we need to test
-    // the fallback to the `standard` variant in this case.
-    shouldMatchEmotionSnapshot(<Alert variant="unknown">Sample Alert</Alert>))
-  ;(['info', 'warning', 'danger', 'success'] as const).forEach(variant => {
-    test(`renders correctly variant ${variant}`, () =>
-      shouldMatchEmotionSnapshot(<Alert variant={variant}>Sample Alert</Alert>))
+  test('renders correctly with buttonText and onButtonClick', () =>
+    shouldMatchEmotionSnapshot(
+      <Alert buttonText="Button" onButtonClick={() => 'ok'}>
+        Sample Alert
+      </Alert>,
+    ))
+
+  test('renders correctly with isClosable and onClose', () =>
+    shouldMatchEmotionSnapshot(
+      <Alert isClosable onClose={() => 'ok'}>
+        Sample Alert
+      </Alert>,
+    ))
+
+  describe(`renders correctly with all variants`, () => {
+    ;(['info', 'warning', 'danger', 'success'] as const).forEach(variant => {
+      test(`renders correctly variant ${variant}`, () =>
+        shouldMatchEmotionSnapshot(
+          <Alert variant={variant}>Sample Alert</Alert>,
+        ))
+    })
   })
 
-  test(`renders correctly unknown theme variant and fallback to defaults`, () =>
-    // @ts-expect-error "unknown" isn't part of the `variant` variant, but we need to test
-    // the fallback to the `theme.colors` colors in this case.
-    shouldMatchEmotionSnapshot(<Alert variant="unknown">Sample Alert</Alert>))
+  test(`should render alert and then close it`, async () => {
+    renderWithTheme(
+      <Alert isClosable onClose={() => 'ok'} data-testid="alert">
+        Sample Alert
+      </Alert>,
+    )
+
+    const alert = screen.getByTestId('alert')
+    expect(alert).toBeVisible()
+
+    const closeButton = screen.getByRole('button')
+    await userEvent.click(closeButton)
+
+    await waitFor(() => {
+      expect(alert).not.toBeVisible()
+    })
+  })
 })