+++ packages/ui/src/components/List/index.tsx
@@ -15,10 +15,10 @@ import {
   useRef,
   useState,
 } from 'react'
-import { orderBy } from '../../utils'
 import type { ComparableType } from '../../utils'
-import { Pagination } from '../Pagination'
+import { orderBy } from '../../utils'
 import type { PaginationProps } from '../Pagination'
+import { Pagination } from '../Pagination'
 import type { UsePaginationReturn } from '../Pagination/usePagination'
 import { usePagination } from '../Pagination/usePagination'
 import { Text } from '../Text'
@@ -478,7 +478,7 @@ const List = forwardRef(
       () =>
         Object.keys(selectableItems).length > 0 &&
         (pagination.pageData as [])
-          .filter(item => selectableItems[item[idKey] as string] === true)
+          .filter(item => selectableItems[item[idKey] as string])
           .every(
             item =>
               rowsState[item[idKey] as string] &&