+++ packages/ui/src/components/SelectInput/__stories__/Playground.stories.tsx
@@ -1,4 +1,3 @@
-import { fireEvent, screen } from '@storybook/testing-library'
 import { SelectInput } from '..'
 import { Template } from './Template.stories'
 
@@ -12,10 +11,6 @@ Playground.args = {
     <SelectInput.Option value="a">Option A</SelectInput.Option>,
     <SelectInput.Option value="b">Option B</SelectInput.Option>,
   ],
-}
-
-Playground.play = () => {
-  fireEvent.click(screen.getByRole('combobox'))
 }
 
 Playground.decorators = [