+++ packages/ui/src/components/VerificationCode/index.tsx
@@ -118,108 +118,111 @@ export const VerificationCode = ({
     }
   }
 
-  const inputOnChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const index = parseInt(String(event.target.dataset?.['id']), 10)
-    let { value } = event.target
-    if (type === 'number') {
-      value = event.target.value.replace(/[^\d]/gi, '')
-    }
-    const newValues = [...values]
+  const inputOnChange =
+    (index: number) => (event: ChangeEvent<HTMLInputElement>) => {
+      let { value } = event.target
+      if (type === 'number') {
+        value = event.target.value.replace(/[^\d]/gi, '')
+      }
+      const newValues = [...values]
 
-    if (
-      value === '' ||
-      (type === 'number' && !new RegExp(event.target.pattern).test(value))
-    ) {
-      newValues[index] = ''
+      if (
+        value === '' ||
+        (type === 'number' && !new RegExp(event.target.pattern).test(value))
+      ) {
+        newValues[index] = ''
+        setValues(newValues)
+
+        return
+      }
+
+      const sanitizedValue = value[0] // in case more than 1 char, we just take the first one
+      newValues[index] = sanitizedValue
       setValues(newValues)
+      const nextIndex = Math.min(index + 1, fields - 1)
+      const next = inputRefs[nextIndex]
 
-      return
+      next?.current?.focus()
+
+      triggerChange(newValues)
     }
 
-    const sanitizedValue = value[0] // in case more than 1 char, we just take the first one
-    newValues[index] = sanitizedValue
-    setValues(newValues)
-    const nextIndex = Math.min(index + 1, fields - 1)
-    const next = inputRefs[nextIndex]
+  const inputOnKeyDown =
+    (index: number): KeyboardEventHandler<HTMLInputElement> =>
+    event => {
+      const prevIndex = index - 1
+      const nextIndex = index + 1
+      const first = inputRefs[0]
+      const last = inputRefs[inputRefs.length - 1]
+      const prev = inputRefs[prevIndex]
+      const next = inputRefs[nextIndex]
+      const vals = [...values]
 
-    next?.current?.focus()
-
-    triggerChange(newValues)
-  }
-
-  const inputOnKeyDown: KeyboardEventHandler<HTMLInputElement> = event => {
-    const index = Number(event.currentTarget.dataset?.['id'])
-    const prevIndex = index - 1
-    const nextIndex = index + 1
-    const first = inputRefs[0]
-    const last = inputRefs[inputRefs.length - 1]
-    const prev = inputRefs[prevIndex]
-    const next = inputRefs[nextIndex]
-    const vals = [...values]
-
-    switch (event.keyCode) {
-      case KEY_CODE.backspace:
-        event.preventDefault()
-        if (values[index]) {
-          vals[index] = ''
-          setValues(vals)
-          triggerChange(vals)
-        } else if (prev) {
-          vals[prevIndex] = ''
+      switch (event.keyCode) {
+        case KEY_CODE.backspace:
+          event.preventDefault()
+          if (values[index]) {
+            vals[index] = ''
+            setValues(vals)
+            triggerChange(vals)
+          } else if (prev) {
+            vals[prevIndex] = ''
+            prev?.current?.focus()
+            setValues(vals)
+            triggerChange(vals)
+          }
+          break
+        case KEY_CODE.left:
+          event.preventDefault()
           prev?.current?.focus()
-          setValues(vals)
-          triggerChange(vals)
-        }
-        break
-      case KEY_CODE.left:
-        event.preventDefault()
-        prev?.current?.focus()
-        break
-      case KEY_CODE.right:
-        event.preventDefault()
-        next?.current?.focus()
-        break
-      case KEY_CODE.up:
-        event.preventDefault()
-        first?.current?.focus()
-        break
-      case KEY_CODE.down:
-        event.preventDefault()
-        last?.current?.focus()
-        break
-      default:
-        break
+          break
+        case KEY_CODE.right:
+          event.preventDefault()
+          next?.current?.focus()
+          break
+        case KEY_CODE.up:
+          event.preventDefault()
+          first?.current?.focus()
+          break
+        case KEY_CODE.down:
+          event.preventDefault()
+          last?.current?.focus()
+          break
+        default:
+          break
+      }
     }
-  }
 
   const inputOnFocus: FocusEventHandler<HTMLInputElement> = event =>
     event.target.select()
 
-  const inputOnPaste: ClipboardEventHandler<HTMLInputElement> = event => {
-    event.preventDefault()
-    const currentIndex = Number(event.currentTarget.dataset?.['id'] as string)
-    const pastedValue = [...event.clipboardData.getData('Text').split('')].map(
-      (copiedValue: string) =>
+  const inputOnPaste =
+    (currentIndex: number): ClipboardEventHandler<HTMLInputElement> =>
+    event => {
+      event.preventDefault()
+      const pastedValue = [
+        ...event.clipboardData.getData('Text').split(''),
+      ].map((copiedValue: string) =>
         // Replace non number char with empty char when type is number
         type === 'number' ? copiedValue.replace(/[^\d]/gi, '') : copiedValue,
-    )
+      )
 
-    // Trim array to avoid array overflow
-    pastedValue.splice(
-      fields - currentIndex < pastedValue.length
-        ? fields - currentIndex
-        : pastedValue.length,
-    )
+      // Trim array to avoid array overflow
+      pastedValue.splice(
+        fields - currentIndex < pastedValue.length
+          ? fields - currentIndex
+          : pastedValue.length,
+      )
 
-    setValues((vals: string[]) => {
-      const newArray = vals.slice()
-      newArray.splice(currentIndex, pastedValue.length, ...pastedValue)
+      setValues((vals: string[]) => {
+        const newArray = vals.slice()
+        newArray.splice(currentIndex, pastedValue.length, ...pastedValue)
 
-      return newArray
-    })
+        return newArray
+      })
 
-    triggerChange(pastedValue)
-  }
+      triggerChange(pastedValue)
+    }
 
   return (
     <div className={className} data-testid={dataTestId}>
@@ -232,13 +235,12 @@ export const VerificationCode = ({
           // eslint-disable-next-line react/no-array-index-key
           key={`${inputId}-${index}`}
           data-testid={index}
-          data-id={index}
           value={value}
           id={inputId ? `${inputId}-${index}` : undefined}
           ref={inputRefs[index]}
-          onChange={inputOnChange}
-          onKeyDown={inputOnKeyDown}
-          onPaste={inputOnPaste}
+          onChange={inputOnChange(index)}
+          onKeyDown={inputOnKeyDown(index)}
+          onPaste={inputOnPaste(index)}
           onFocus={inputOnFocus}
           disabled={disabled}
           required={required}